+++ src/Feliz.CompilerPlugins/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Relax validation of record props defined along the react component to allow lower cased record types #463 #666 by @melanore
+
 ## 3.0.0-rc.1 - 2025-09-18
 
 ### Added